requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pyPDF2
faiss-cpu
langchain_google_genai
pydub
sse_starlette
requests
uvicorn
fastapi
langchain_google_genai
langchain_core
chroma
